requirements.txt
pydantic==2.5.3
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
fastapi==0.109.0
tortoise-orm[asyncpg]==0.20.0
uvicorn==0.27.0
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.26.0
anyio==4.2.0